tests/wpt/web-platform-tests/css/css-position/parsing/inset-after-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing inset-after with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-inset-after">
<meta name="assert" content="inset-after supports the full grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("inset-after", "auto");

test_valid_value("inset-after", "-10px");
test_valid_value("inset-after", "-20%");
test_valid_value("inset-after", "calc(2em + 3ex)");
</script>
</body>
</html>
